Ahead Java A1 Q2
Define an array of size N, the size should not exceed 20. Input N integers into the array. Find the sum of elements in even indices and product of elements of odd indices. There should be atleast two input elements to the array. Hint: index 0 can be considered as even index.

Input Format

Input an integer for N In the next line read N integers in a single line space separated.

Constraints

The value of N and the integers to the array should be a positive integer

Output Format

If the value of N is not between 2 and 20, print INVALID. If any of the integers entered into the array is neagtive, print INVALID. Otherwise Print the Sum and Product results in two separate lines.

Sample Input 0

4
1 2 3 4
Sample Output 0

4
8
Sample Input 1

8
1 -2
Sample Output 1

INVALID