Shapes - Class Inheritance Easy
Shapes - Class Inheritance
Easy

Time Limit: 2 sec
Memory Limit: 128000 kB
Problem Statement
Complete the Main class. Implement the getArea() and getPerimeter() methods for the child classes Parallelogram, Rhombus, Rectangle, and Square having Abstract class Quadrilateral as a parent class.


Do not change the name of any class
Input
The first line of input contains side1, side2 and height of the parallelogram
The second line of input contains side and height of rhombus
The third line of input contains length and height of the rectangle
The fourth line of input contains side of square

If the length of any side or height for any shape is negative, then print Length of a side cannot be negative. Please Enter a positive integer.
Output
"Perimeter of Parallelogram is " + parallelogram.getPerimeter() +" and Area of Parallelogram is " + parallelogram.getArea()
"Perimeter of Rhombus is " + rhombus.getPerimeter() +" and Area of Rhombus is " + rhombus.getArea()
"Perimeter of Rectangle is " + rectangle.getPerimeter() +" and Area of Rectangle is " + rectangle.getArea()
"Perimeter of Square is " + square.getPerimeter()+ " and Area of Square is " + square.getArea()



Example
Sample Input 1:
1.0 2.0 3.0
7.0 5.0
4.0 2.0
6.0

Sample Output 2:
Perimeter of Parallelogram is 6.0 and Area of Parallelogram is 3.0

Perimeter of Rhombus is 28.0 and Area of Rhombus is 35.0

Perimeter of Rectangle is 12.0 and Area of Rectangle is 8.0

Perimeter of Square is 24.0 and Area of Square is 36.0


Sample Input 2:
3.0 9.0 7.0
6.0 4.0
-1.0 5.0
8.0

Sample Output 2:
Length of a side cannot be negative. Please Enter a positive integer