Operations on ArrayList Easy
Operations on ArrayList
Easy

Problem Statement
Given an arraylist of integers. The task is to perform operations on ArrayList according to the queries and print output for the required queries.

Note: use Collections.sort() to sort the list in natural order, Collections.reverseOrder() to arrange the elements in reverse, get() to return the element at index of the list, linear search for finding specified element in the list. If multiple positions exist, print the smallest one.
Input
User Task:
Since this is a functional problem, you don’t have to worry about the input, you just have to complete the functions:-
insert() That takes the arraylist and the integer to be inserted as a parameter
IncOrder() That takes the arraylist as a parameter.
Search()That takes the arraylist and the integer to be find as a parameter.
DecOrder() That takes arraylist as a parameter.

Constraints:-
1 < = N < = 1000
1 < = Arr[i] < = 100000
Output
Print the index of element if found else print -1 in Search function add a white-space after every output.
Example
Sample Input:-
6
a 2 a 3 a 4 a 2 i s 2

Sample output:-
0

Sample Input:-
5
a 2 a 3 a 3 d s 5

Sample Output:-
-1

Explanation:
Testcase 1: After inserting elements, list is (2, 3, 4, 2) and arranging them in ascending order, we have list as (2, 2, 3, 4). Now, serching for 2, first occurence is at 0th index.

Testcase 2: After inserting elements, list is (2, 3, 3) and arranging them in descending order, we have list as (3, 3, 2). Finding 5 gives -1, as 5 is not present.