Maximum of minimum of every window size  Hard
Maximum of minimum of every window size
Hard

Time Limit: 2 sec
Memory Limit: 128000 kB
Problem Statement
Given an integer array A[] of size N. For any subarray size i (1 <= i <= N), you have to do the following:
Write all subarrays of size i from array A on a piece of paper.
Now, replace each of these subarrays with their minimum element.
Now, find the maximum value among all these minimums.

Print this maximum value for all subarray sizes k (1 <= k <= N).
Input
The first line contains an integer N denoting the size of the array. The second line contains N space-separated integers A1, A2, ..., AN denoting the elements of the array.

Constraint:-
1 <= N <= 10^5
1 <= A[i] <= 10^6
Output
Print the array of numbers of size N for each of the considered window size 1, 2, ..., N respectively.
Example
Sample Input:
7
10 20 30 50 10 70 30

Sample Output:
70 30 20 10 10 10 10

Explanation:
Testcase 1:
First element in output indicates maximum of minimums of all windows of size 1. Minimums of windows of size 1 are {10}, {20}, {30}, {50}, {10}, {70} and {30}. Maximum of these minimums is 70.
Second element in output indicates maximum of minimums of all windows of size 2. Minimums of windows of size 2 are {10}, {20}, {30}, {10}, {10}, and {30}. Maximum of these minimums is 30.
Third element in output indicates maximum of minimums of all windows of size 3. Minimums of windows of size 3 are {10}, {20}, {10}, {10} and {10}. Maximum of these minimums is 20.
Similarly other elements of output are computed.