The Social Network (Public Contest: June 2022)
The Social Network (Public Contest: June 2022)
Time Limit: 1 sec
Memory Limit: 128000 kB
Problem Statement
There are N persons standing in front of the entrance of a hotel with M rooms. Initially, all the rooms are empty and the people are standing in a queue to enter the hotel. We know that each of the N persons is either an Introvert or otherwise Social. Now one by one, they will enter the hotel. If the person is an Introvert, he will go to the least filled room and stay there till next week. If there are multiple least filled rooms, then the person will go to any of those with equal probability.

On the other hand, as you may have guessed, a Social person will go to the most filled room. If there are multiple most filled rooms, then the person will go to any of those with equal probability. Once all the N persons have entered their rooms, you need to find for each person, the expected number of people in their room.

It can be shown that the expected value can be written in the form of p/q, where p and q are coprime integers and gcd(q, 998244353) = 1. You task is to print the value p×q-1 modulo 998244353, where q-1 is the modular inverse of q under modulo 998244353.
Input
The first line consists of two space-separated integers N and M.

The second line consists of a string S of size N, consisting only of characters 'I' and 'S'. The ith character of the string denotes the type of the person standing at the ith position from the front of the queue. If it is 'I', then the person is an Introvert, otherwise he is Social.

Constraints:
1 ≤ N ≤ 105
1 ≤ M ≤ 105
Output
Print N lines. The ith line should a contain a single integer – the expected number of persons in the room where the ith person is staying.
Example
Sample Input 1:
5 2
ISIIS

Sample Output 1:
499122179
499122179
499122179
499122179
3

Sample Explanation 1:
The first person can go to either room 1 or room 2. The second person will go and join the first person in their room. Since the 3rd person is introvert, they will go to the other empty room. The fourth person is also an introvert, so he will join the third person, as that room is least filled. The last person can join any of the two rooms with equal probability, since they both contain the same number of people.