Add two numbers Medium
Add two numbers
Medium

Problem Statement
Two numbers are represented in Linked List such that each digit corresponds to a node in linked list. Your task is to add these two numbers and return the sum in a linked list.

Note:-Linked list representaion of a number is from left to right i.e if the number is 123 than in linked list it is represented as 3->2->1
Input
User Task:
Since this will be a functional problem, you don't have to take input. You just have to complete the function addNumber() that takes head nodes of both the linked lists as parameter.

Constraints:
1 <=numbers<=10^1000
Output
Return the head of modified linked list.
Example
Sample Input:-
1234 45643


Sample Output:-
46877